feat: Extract block placement and digging pipelines over a world interface
BlockPlacementListener and PlayerActionListener's digging half move into
a public BlockActions parameterized by a World: block access, the
placement write, the break, chunk/read-only checks, and the placement
entity scan, defaulting to the player's instance. The stock listeners
delegate with the instance-backed world, behavior unchanged.

Virtual worlds and block overlays can now run the same pipeline (target
resolution, events, acknowledgements, client resyncs) against their own
storage instead of forking the listeners packet by packet.
